Progreso en la navegación y limpieza
Queda que cambiar de botón en las escenas de scroll le afecte a este. Además de mover el input del portal a PlayerInputComponent.

+++ Game/Game/src/data/CardData.h
@@ -10,7 +10,7 @@ struct CardData {
     int damage;
     int maxUses;
     int mana;
-    int downTime;
+    float downTime;
     Texture* texture;
     CardTime time;
 };
